Scrud Person done - all fixed

+++ src/main/java/com/safetynet/alerts/service/PersonService.java
@@ -6,44 +6,64 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.safetynet.alerts.model.Person;
 import com.safetynet.alerts.repository.PersonRepository;
 
 @Service
-public class PersonService implements IPersonService {
+public class PersonService implements IPersonService
+{
+	@Autowired
+	private PersonRepository personRepository;
 	
 	private static Logger logger = LogManager.getLogger("PersonService");
 	
-	@Autowired
-	private PersonRepository personRepository;
-    
-	public List<Person> getAllPersons(){
-		logger.info("Person list found");
+	public List<Person> getAllPersons()
+	{
 		return personRepository.getAllPerson();
 	}
 	
 	@Override
-    public Person getPersonByName(String firstName, String lastName) {
-		logger.info("Person : {} {} is found", firstName,lastName);
-    	return personRepository.getPersonByName(firstName,lastName);
-    }
-	
-	@Override
-	public Person addPerson(Person person) {
-		logger.info("Person added : {}", person);
+	public Person addPerson(Person person) 
+	{
 		return personRepository.addPerson(person);
 	}
 	
 	@Override
-	public Person updatePerson(Person person) {
-		logger.info("New person info : {} {} is update", person.getFirstName(),person.getLastName());
-		return personRepository.updatePerson(person);
+    public Person getPersonByName(String firstName, String lastName) 
+	{
+		for (Person personToFind : personRepository.getAllPerson()) 
+		{
+			if(personToFind.getFirstName().equals(firstName) && 
+					personToFind.getLastName().equals(lastName)) 
+			{
+				
+				return personRepository.getPersonByName(firstName, lastName);
+			}
+		}
+		logger.info("No match! Person not Found.");
+		return null;
+    }
+	
+	@Override
+	public Person updatePerson(Person person) 
+	{
+		for (Person personToUpdate : personRepository.getAllPerson()) 
+		{
+			if (personToUpdate.getFirstName().equals(person.getFirstName()) &&
+					personToUpdate.getLastName().equals(person.getLastName())) 
+			{
+				return personRepository.updatePerson(person);
+			}
+		}
+		return null;
 	}
 	
 	@Override
-	public void deletePerson(Person person) {
-		logger.info("Person : {} {} is delete", person.getFirstName(),person.getLastName());
-		personRepository.deletePerson(person);
+	public void deletePerson(Person person) 
+	{
+		if (personRepository.getAllPerson().contains(person)) 
+		{
+			personRepository.deletePerson(person);
+		}
 	}
 }
